feat: 3단계 피드백 반영 | Sections -> 일급컬렉션
