+++ src/main/java/nextstep/subway/domain/Line.java
@@ -1,12 +1,7 @@
 package nextstep.subway.domain;
 
-import nextstep.subway.applicaion.exception.NotFoundException;
-
 import javax.persistence.*;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.stream.Collectors;
 
 @Entity
 public class Line extends BaseEntity {
@@ -18,13 +13,13 @@ public class Line extends BaseEntity {
     private String name;
     private String color;
 
-    @OneToMany(mappedBy = "line", cascade = {CascadeType.PERSIST, CascadeType.MERGE}, orphanRemoval = true)
-    private List<Section> sections = new ArrayList<>();
+    @Embedded
+    private Sections sections = new Sections();
 
     public Line() {
     }
 
-    public Line(Long id, String name, String color, List<Section> sections) {
+    public Line(Long id, String name, String color, Sections sections) {
         this.id = id;
         this.name = name;
         this.color = color;
@@ -43,7 +38,7 @@ public class Line extends BaseEntity {
         return color;
     }
 
-    public List<Section> getSections() {
+    public Sections getSections() {
         return sections;
     }
 
@@ -52,52 +47,12 @@ public class Line extends BaseEntity {
         this.color = color;
     }
 
-    public void addSection(Station upStation, Station downStation, int distance) {
-        validateStationInSection(downStation, upStation);
-        this.sections.add(new Section(this, upStation, downStation, distance));
+    public void deleteLastSection(Station foundStation) {
+        this.sections.deleteLastSection(foundStation);
     }
 
-
-    public Section getLastSection() {
-        final int lastIndex = this.sections.size() - 1;
-        return this.sections.get(lastIndex);
-    }
-
-    public void validateStationInSection(Station downStation, Station upStation) {
-        if (!this.sections.isEmpty()) {
-            checkUpStation(upStation);
-            checkDownStation(downStation);
-        }
-    }
-
-    private void checkUpStation(Station upStation) {
-        final Section lastSection = getLastSection();
-        if (!upStation.equals(lastSection.getDownStation())) {
-            throw new IllegalArgumentException("등록할 상행종점역은 노선의 하행종점역이어야 합니다.");
-        }
-    }
-
-    private void checkDownStation(Station downStation) {
-        final List<Station> stations = this.sections.stream()
-                .flatMap(section -> section.getAllStation().stream())
-                .distinct()
-                .collect(Collectors.toList());
-        if (stations.contains(downStation)) {
-            throw new IllegalArgumentException("등록할 하행종점역은 노선에 등록되지 않은 역만 가능합니다.");
-        }
-    }
-
-    public void deleteLastSection(Station station) {
-        if (this.sections.size() <= 1) {
-            throw new IllegalArgumentException("지하철 구간이 1개인 경우 구간을 제거할 수 없습니다.");
-        }
-
-        final Section lastSection = getLastSection();
-        if (!lastSection.getDownStation().equals(station)) {
-            throw new IllegalArgumentException("노선에 등록된 역(하행종점역)만 제거 가능합니다.");
-        }
-
-        this.sections.remove(lastSection);
+    public void addSection(Station upStation, Station downStation, Integer distance) {
+        this.sections.addSection(this, upStation, downStation, distance);
     }
 
 
@@ -105,7 +60,7 @@ public class Line extends BaseEntity {
         private Long id;
         private String name;
         private String color;
-        private List<Section> sections = new ArrayList<>();
+        private Sections sections = new Sections();
         private LocalDateTime createdDate;
         private LocalDateTime modifiedDate;
 
@@ -124,7 +79,7 @@ public class Line extends BaseEntity {
             return this;
         }
 
-        public Builder sections(List<Section> sections) {
+        public Builder sections(Sections sections) {
             this.sections = sections;
             return this;
         }
